fix: logic
